PackageSet
PackageSet구

+++ src/main/java/net/sourceforge/pmd/lang/java/rule/stanly/metrics/PackageSetMetric.java
@@ -59,7 +59,7 @@ public class PackageSetMetric{
 	public int getFat() {
 		return Fat;
 	}
-	public void setFat(float fAT) {
+	public void setFat(int fAT) {
 		Fat = fAT;
 	}
 	public float getTangled() {
